Add contact section time module
